Use isolated
We moved to use pypi to install rally, but we didn't remove
the pins, nor use isolated.

This is a problem because the repo_build could have built a series
of packages which are common to a different role.
In this case, ansible gets pulled by ceph-ansible, and by rally.
Indeed, ceph-ansible has recently added requirements into
their repository [3].

The ansible version pulled by ceph-ansible is in [1], which
would resolve this day to ansible 2.4.4.0.
The ansible wheel built would not match the one pinned in the
rally role.

The repo server serves the packages locally if a local package
exists for the same name, and relies on pypi otherwise.

When the repo server builds the virtualenv, the venv creation
for rally fails, as the repo server wants to serve the ceph-ansible
built version of ansible, and not the 2.4.3.0 asked by rally.

The best way to fix this is with a multiple series of patches:
1 Ensure the repo server can serve all the versions of a package,
  by pointing to upstream if a specific version was requested
2 Temporarily ensure the rally installation is ignoring the
  constraints of other packages, and installing the packages
  without being limited by the repository server
3 Remove the unnecessary pin, to follow rally's recommended
  version.
4 Ensure the rally installation version is installed with
  the matching upstream repo constraints [2].
5 Ensure ceph-ansible's requirements are not installed on
  the destination node, nor built into the repo-server.

This handles the 2) case, and starts the 3) work.
A follow-up patch should be included to ensure 4) is done,
ensuring the ansible pin setup in this patch can be removed.

[1]: https://github.com/ceph/ceph-ansible/blob/f68936ca7e7f556c8d8cee8b2c4565a3c94f72f9/requirements.txt
[2]: https://github.com/openstack/rally/blob/0.12.1/upper-constraints.txt
[3]: https://github.com/ceph/ceph-ansible/commit/ef48ed4e5af10f93f7bcf6d19ab48143554f2038#diff-b4ef698db8ca845e5845c4618278f29a

Change-Id: I476bd2ae645aa5e0628924b631c4cc4cba4c93d8